fix(database): preserve fractional calibration values

+++ install/database/scale_db.sql
@@ -53,7 +53,7 @@ CREATE TABLE calibration_points (
     adc_code   INTEGER      NOT NULL,
     mass       NUMERIC(6,2) NOT NULL CHECK (mass >= 0 AND mass <= 150),
     is_active  BOOLEAN      NOT NULL DEFAULT TRUE,
-    calibration_value INTEGER      NOT NULL,
+    calibration_value NUMERIC(12,3) NOT NULL,
     created_at TIMESTAMPTZ  NOT NULL DEFAULT NOW(),
     deleted_at TIMESTAMPTZ
 );
